third_party/blink/web_tests/http/tests/cookies/partitioned-cookies/partitioned-cookies-in-parallel-frames.https.html

<!doctype html>
<head>
<meta charset="utf-8"/>
<meta name="timeout" content="long">
<meta name="help" href="https://github.com/WICG/CHIPS#chips-cookies-having-independent-partitioned-state">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/cookies/resources/testharness-helpers.js"></script>
<title>Test partitioned cookie behavior in parallel embeds</title>
</head>
<body>
<script>

function cookiePresent(cookieName, frame) {
  return frame.contentDocument.cookie.includes(cookieName) &&
      frame.contentWindow.location.href.includes(cookieName);
}

function iframeHost(iframe) {
  iframeUrl = new URL(iframe.src);
  return iframeUrl.hostname;
}

function ObserveLoadEvents(frames, triggerEvents) {
  const promises = frames.map((frame) => new Promise((resolve) => {
                                frame.onload = resolve;
                              }));
  triggerEvents();
  return Promise.all(promises);
}

promise_test(async () => {
  // Add partitioned cookie to top-level site.
  const partitionedCookie = 'ancestor=chain';
  const partitionedCookieAttributes =
      '; Secure; Path=/; SameSite=None; Partitioned';
  const partitionedCookieLine = partitionedCookie + partitionedCookieAttributes;

  document.cookie = partitionedCookieLine;
  assert_true(document.cookie.includes(partitionedCookie));

  const sameSiteHost = window.location.hostname;
  const resourceDir = '/cookies/partitioned-cookies/resources/';
  const crossSiteUrl = new URL(
      resourceDir + 'ancestor-chain-empty-embed.html',
      `https://${TEST_HOST}:${window.location.port}`);
  const sameSiteUrl = new URL(
      resourceDir + 'ancestor-chain-empty-embed.html',
      `https://${sameSiteHost}:${window.location.port}`);

  // Create two cross-site iframes and wait for them to load.
  const iframe = document.createElement('iframe');
  const parallelFrame = document.createElement('iframe');

  // Wait for initial load of iframes that occurs when the iframe is added to
  // the document.
  // https://html.spec.whatwg.org/multipage/iframe-embed-object.html#the-iframe-element:process-the-iframe-attributes
  await ObserveLoadEvents([iframe, parallelFrame], () => {
    document.body.appendChild(iframe);
    document.body.appendChild(parallelFrame);
  });

  // Wait for load that occurs after changing src attribute.
  await ObserveLoadEvents([iframe, parallelFrame], () => {
    iframe.src = crossSiteUrl.href;
    parallelFrame.src = crossSiteUrl.href;
  });

  // Confirm that both iframes are cross-site to the top-level document and
  // same-site to each other.
  assert_not_equals(sameSiteHost, iframeHost(iframe));
  assert_not_equals(sameSiteHost, iframeHost(parallelFrame));
  assert_equals(iframeHost(iframe), iframeHost(parallelFrame));

  // Initiate the redirect from cross-site to same-site through
  // redirect-and-append-cookie-header.php
  const redirectUrl = resourceDir +
      'redirect-and-append-cookie-header.php?location=' + sameSiteUrl.href;

  await ObserveLoadEvents([iframe, parallelFrame], () => {
    iframe.src = redirectUrl;
    parallelFrame.src = redirectUrl;
  });

  // Confirm that the iframes are now same-site to document.
  assert_equals(sameSiteHost, iframeHost(iframe));
  assert_equals(sameSiteHost, iframeHost(parallelFrame));

  // Confirm that the cookie was in http header of the redirect requests and in
  // document.
  assert_true(cookiePresent(partitionedCookie, iframe));
  assert_true(cookiePresent(partitionedCookie, parallelFrame));
}, 'Partitioned cookies are sent in parallel embedded cross-site to same-site redirects');

promise_test(async t => {
  // Add partitioned cookie to top-level site.
  const partitionedCookie = 'ancestor=chain';
  const partitionedCookieAttributes =
      '; Secure; Path=/; SameSite=None; Partitioned';
  const partitionedCookieLine = partitionedCookie + partitionedCookieAttributes;

  document.cookie = partitionedCookieLine;

  assert_true(document.cookie.includes(partitionedCookie));

  const resourceDir = '/cookies/partitioned-cookies/resources/';
  const sameSiteRedirectUrl = resourceDir +
      'redirect-and-append-cookie-header.php?location=' +
      'ancestor-chain-empty-embed.html';
  const crossSiteUrl = new URL(
      resourceDir + 'ancestor-chain-cross-site-parallel-embed.html',
      `https://${TEST_HOST}:${window.location.port}`);
  const crossSiteRedirectUrl = new URL(
      resourceDir +
          'redirect-and-append-cookie-header.php?location=' + crossSiteUrl,
      `https://${TEST_HOST}:${window.location.port}`);

  const iframe = document.createElement('iframe');
  const parallelFrame = document.createElement('iframe');

  // Wait for initial load of iframes that occurs when the iframe is added to
  // the document.
  // https://html.spec.whatwg.org/multipage/iframe-embed-object.html#the-iframe-element:process-the-iframe-attributes
  await ObserveLoadEvents([iframe, parallelFrame], () => {
    document.body.appendChild(iframe);
    document.body.appendChild(parallelFrame);
  });

  // Redirect the page back to the same-site to allow the
  // redirect-and-append-cookie-header.php file to append the cookie header to
  // the url.
  await ObserveLoadEvents([iframe, parallelFrame], () => {
    iframe.src = sameSiteRedirectUrl;
    parallelFrame.src = sameSiteRedirectUrl;
  });

  // Confirm that the iframe is same-site to the top-level site.
  assert_equals(window.location.hostname, iframeHost(iframe));
  assert_equals(window.location.hostname, iframeHost(parallelFrame));


  // Confirm that the partitioned cookie was in http header of the redirect
  // requests and in document.
  assert_true(cookiePresent(partitionedCookie, iframe));
  assert_true(cookiePresent(partitionedCookie, parallelFrame));

  const cookiePresences = [];
  const allMessagesReceived = new Promise((resolve) => {
    // Add listener that will respond to the messages sent by the cross-site
    // iframes.
    window.addEventListener('message', function(event) {
      const data = event.data;
      cookiePresences.push(data.isCookiePresent);
      if (cookiePresences.length == 2) {
        resolve();
      }
    });
  });

  // Change src to redirect the iframes to be cross-site and wait for
  // them to load.
  await ObserveLoadEvents([iframe, parallelFrame], () => {
    iframe.src = crossSiteRedirectUrl;
    parallelFrame.src = crossSiteRedirectUrl;
  });

  // Confirm that the iframe is cross-site to the top-level site and same-site
  // to each other.
  assert_not_equals(window.location.hostname, iframeHost(iframe));
  assert_not_equals(window.location.hostname, iframeHost(parallelFrame));
  assert_equals(iframeHost(iframe), iframeHost(parallelFrame));

  // Wait for both iframes to send messages to the top-level site indicating
  // if the partitioned cookie is present.
  await allMessagesReceived;

  // Confirm that the cookie is not present in either message sent by the
  // cross-site iframes.
  assert_array_equals(cookiePresences, [false, false]);
}, 'Partitioned cookies are not sent in parallel embedded same-site to cross-site redirects');

</script>
</body>
